refactor: Organize the Code in folders
