+++ Main.qml
@@ -3,6 +3,7 @@ import QtQuick.Controls
 import QtQuick.Controls.Material
 import QtQuick.Layouts
 import DialSome
+import "ui"
 
 
 ApplicationWindow {
