+++ android/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.github.biltudas1.dialsome" android:installLocation="auto" android:versionCode="6" android:versionName="0.4.1-alpha">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.github.biltudas1.dialsome" android:installLocation="auto" android:versionCode="7" android:versionName="0.4.2-alpha">
     <!-- %%INSERT_PERMISSIONS -->
     <!-- %%INSERT_FEATURES -->
         <uses-sdk android:minSdkVersion="28" android:targetSdkVersion="36"/>
